+++ R/test.R
@@ -3,6 +3,8 @@
 #' Creates `tests/testthat/`, `tests/testthat.R`, and adds the testthat package
 #' to the Suggests field. Learn more in <https://r-pkgs.org/tests.html>
 #'
+#' @param edition testthat edition to use. Defaults to the latest edition,
+#'   i.e. the major version number of the currently installed testthat.
 #' @seealso [use_test()] to create individual test files
 #' @export
 #' @examples
@@ -13,8 +15,8 @@
 #'
 #' use_test("something-management")
 #' }
-use_testthat <- function() {
-  use_testthat_impl()
+use_testthat <- function(edition = NULL) {
+  use_testthat_impl(edition)
 
   ui_todo(
     "Call {ui_code('use_test()')} to initialize a basic test file and open it \\
@@ -22,7 +24,7 @@ use_testthat <- function() {
   )
 }
 
-use_testthat_impl <- function() {
+use_testthat_impl <- function(edition = NULL) {
   check_installed("testthat")
   if (utils::packageVersion("testthat") < "2.1.0") {
     ui_stop("testthat 2.1.0 or greater needed. Please install before re-trying")
@@ -30,6 +32,9 @@ use_testthat_impl <- function() {
 
   if (is_package()) {
     use_dependency("testthat", "Suggests")
+
+    edition <- check_edition(edition)
+    use_description_field("Config/testthat/edition", edition)
   }
 
   use_directory(path("tests", "testthat"))
@@ -38,6 +43,22 @@ use_testthat_impl <- function() {
     save_as = path("tests", "testthat.R"),
     data = list(name = project_name())
   )
+}
+
+check_edition <- function(edition = NULL) {
+  testthat_version <- packageVersion("testthat")[[1, 1]]
+  if (is.null(edition)) {
+    testthat_version
+  } else {
+    if (!is.numeric(edition) || length(edition) != 1) {
+      ui_stop("`edition` must be a single number")
+    }
+    if (edition > testthat_version) {
+      vers <- packageVersion("testthat")
+      ui_stop("`edition` ({edition}) not available in installed testthat ({vers}")
+    }
+    as.integer(edition)
+  }
 }